refactor: modularize scripts
