+++ .eslintrc.json
@@ -6,7 +6,7 @@
   "extends": "eslint:recommended",
   "parserOptions": {
     "ecmaVersion": 12,
-    "sourceType": "script"
+    "sourceType": "module"
   },
   "rules": {
     "indent": ["error", 2],
