+++ index.html
@@ -386,9 +386,6 @@
 
   <!-- Main JavaScript modules -->
   <script type="module" src="scripts.js"></script>
-  <script type="module" src="js/navigation.js"></script>
-  <script type="module" src="js/carousel.js"></script>
-  <script type="module" src="js/forms.js"></script>
   <!-- SVG Icons -->
   <svg xmlns="http://www.w3.org/2000/svg" style="display:none">
     <symbol id="icon-sun" viewBox="0 0 24 24">
