+++ js/theme.js
@@ -1,0 +1,17 @@
+export function initTheme() {
+  const themeToggleBtn = document.getElementById('theme-toggle');
+  if (!themeToggleBtn) return;
+  const prefersDark = window.matchMedia('(prefers-color-scheme: dark)').matches;
+  const savedTheme = localStorage.getItem('theme') || (prefersDark ? 'dark' : 'light');
+  document.documentElement.setAttribute('data-theme', savedTheme);
+  themeToggleBtn.setAttribute('aria-checked', savedTheme === 'dark' ? 'true' : 'false');
+  themeToggleBtn.addEventListener('click', () => {
+    const current = document.documentElement.getAttribute('data-theme');
+    const next = current === 'dark' ? 'light' : 'dark';
+    document.documentElement.setAttribute('data-theme', next);
+    localStorage.setItem('theme', next);
+    themeToggleBtn.setAttribute('aria-checked', next === 'dark' ? 'true' : 'false');
+  });
+}
+
+export default initTheme;
