+++ scripts.js
@@ -4,35 +4,42 @@
 
 
 
+import { initNavigation } from './js/navigation.js';
+import { initCarousel } from './js/carousel.js';
+import { initForms } from './js/forms.js';
+import { initTheme } from './js/theme.js';
+
 /* =============================================================================
    PROJECT FILTERING
    ========================================================================== */
-const filterButtons = document.querySelectorAll('.filter-btn');
-const projectCards = document.querySelectorAll('.project-card');
+export function initProjectFiltering() {
+  const filterButtons = document.querySelectorAll('.filter-btn');
+  const projectCards = document.querySelectorAll('.project-card');
 
-filterButtons.forEach(btn => {
-  btn.addEventListener('click', () => {
-    // Toggle active class
-    filterButtons.forEach(b => b.classList.remove('filter-btn--active'));
-    btn.classList.add('filter-btn--active');
-    const filter = btn.getAttribute('data-filter');
-    projectCards.forEach(card => {
-      const category = card.getAttribute('data-category');
-      if (filter === 'all' || category === filter) {
-        card.style.display = 'block';
-        card.classList.remove('filtered-out');
-      } else {
-        card.style.display = 'none';
-        card.classList.add('filtered-out');
-      }
+  filterButtons.forEach(btn => {
+    btn.addEventListener('click', () => {
+      // Toggle active class
+      filterButtons.forEach(b => b.classList.remove('filter-btn--active'));
+      btn.classList.add('filter-btn--active');
+      const filter = btn.getAttribute('data-filter');
+      projectCards.forEach(card => {
+        const category = card.getAttribute('data-category');
+        if (filter === 'all' || category === filter) {
+          card.style.display = 'block';
+          card.classList.remove('filtered-out');
+        } else {
+          card.style.display = 'none';
+          card.classList.add('filtered-out');
+        }
+      });
     });
   });
-});
+}
 
 /* =============================================================================
    SCROLL-ON-LOAD ANIMATIONS (IntersectionObserver)
    ========================================================================== */
-function initScrollAnimations() {
+export function initScrollAnimations() {
   const elements = document.querySelectorAll('.animate-on-scroll');
   const observerOptions = {
     root: null,
@@ -54,7 +61,7 @@ function initScrollAnimations() {
 /* =============================================================================
    LAZY-LOADING IMAGES & BGs (IntersectionObserver)
    ========================================================================== */
-function initLazyLoad() {
+export function initLazyLoad() {
   const lazyImages = document.querySelectorAll('img[loading="lazy"]');
   const lazyBackgrounds = document.querySelectorAll('[data-bg]');
   const observerOptions = {
@@ -101,7 +108,7 @@ function getCookie(name) {
   }, '');
 }
 
-function checkCookieConsent() {
+export function checkCookieConsent() {
   if (getCookie('cookieConsent') !== 'true') {
     cookieBanner.classList.add('show');
   }
@@ -112,26 +119,16 @@ acceptCookiesBtn.addEventListener('click', () => {
   cookieBanner.classList.remove('show');
 });
 
-document.addEventListener('DOMContentLoaded', () => {
-  checkCookieConsent();
-});
-
 /* =============================================================================
    INITIALIZATIONS ON DOM CONTENT LOADED
    ========================================================================== */
 document.addEventListener('DOMContentLoaded', () => {
+  initNavigation();
+  initCarousel();
+  initForms();
+  initTheme();
+  initProjectFiltering();
   initScrollAnimations();
   initLazyLoad();
-  const themeToggleBtn = document.getElementById('theme-toggle');
-  const prefersDark = window.matchMedia('(prefers-color-scheme: dark)').matches;
-  const savedTheme = localStorage.getItem('theme') || (prefersDark ? 'dark' : 'light');
-  document.documentElement.setAttribute('data-theme', savedTheme);
-  themeToggleBtn.setAttribute('aria-checked', savedTheme === 'dark' ? 'true' : 'false');
-  themeToggleBtn.addEventListener('click', () => {
-    const current = document.documentElement.getAttribute('data-theme');
-    const next = current === 'dark' ? 'light' : 'dark';
-    document.documentElement.setAttribute('data-theme', next);
-    localStorage.setItem('theme', next);
-    themeToggleBtn.setAttribute('aria-checked', next === 'dark' ? 'true' : 'false');
-  });
+  checkCookieConsent();
 });
